feat(analytics): add Google Analytics and supply env vars through `scripts/bin.ts` (#8)
Fixes https://github.com/imballinst/how-long/issues/7. This PR adds GA to the Astro project.

The GA script is applied in the `base/Head.astro`, which then we call it in the React components. This is possible in TypeScript because we have `@types/gtag.js` in our `devDependencies`. Hence, there is no hacky stuff or TypeScript error whatsoever.

In addition, since we need `PUBLIC_*` environment variables to indicate production (and to flag the analytics off in dev environments), I have created a initializer script that will fully forward the command passed to `packages/app` workspace. For example, `yarn start` will do `astro start`, in addition to the `PUBLIC_*` environment variables.