feat(date): add generate/last update dates in all views (#9)
Fixes https://github.com/imballinst/how-long/issues/4. This PR adds the text "This page was last updated at ...". This date string is generated at build time.

This also changes the `collection.json` to `items.json`, as well as changing its data structure from an array to an object of `{ date, items }`, in which `items` contains the previous array. The `date` field contains the generate date.